+++ log/warn.ts
@@ -19,3 +19,29 @@ export function warn<T>(
   }
   return getLogger("default").warn(msg, ...args);
 }
+
+/**
+ * @deprecated (will be removed after 0.214.0) Use {@linkcode warn} instead.
+ *
+ * Log with warning level, using default logger.
+ */
+export function warning<T>(msg: () => T, ...args: unknown[]): T | undefined;
+/**
+ * @deprecated (will be removed after 0.214.0) Use {@linkcode warn} instead.
+ *
+ * Log with warning level, using default logger.
+ */
+export function warning<T>(
+  msg: T extends GenericFunction ? never : T,
+  ...args: unknown[]
+): T;
+export function warning<T>(
+  msg: (T extends GenericFunction ? never : T) | (() => T),
+  ...args: unknown[]
+): T | undefined {
+  // Assist TS compiler with pass-through generic type
+  if (msg instanceof Function) {
+    return warn(msg, ...args);
+  }
+  return warn(msg, ...args);
+}
